proper test locations
